Revert s3-proxy caching back to 24h
There is a problem with root pages that default to index.htm. If cache does not exist, the request fails with an error from AWS:
---
The request signature we calculated does not match the signature you provided. Check your key and signing method.
---

To work around this, I put the cache to 24 hours again, and installed a cronjob that requests https://aboutbears.ostankin.net/index.htm once a day to prevent requests to https://aboutbears.ostankin.net from failing.